requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
xgboost==2.0.3
requests
apscheduler
python-dotenv
pytest
httpx
